Users routes reworked, using fs.promises
