+++ README.md
@@ -1,6 +1,6 @@
 # Проектная работа 12, Мун Дмитрий
 ## Актуальная версия проекта
-0.0.3
+0.0.4
 ## Описание проекта
 Проектная работа 12 из курса веб-разработки [Яндекс.Практикума](https://praktikum.yandex.ru/)
 ## Использованные технологии
@@ -14,7 +14,7 @@
 `git clone https://github.com/mun-dmitry/sprint-12__yaPraktikum`
 2. Установите зависимости
 `npm install`
-3. 
+3. Работайте с проектом
 - Запуск сервера
 `npm run start`
 - Запуск сервера в режиме разработки
