+++ routes/router.js
@@ -1,6 +1,11 @@
 const router = require('express').Router();
-const { doesUserExist, sendUser, sendUsers, noSuchAddress } = require('./routes.js');
 const sendCards = require('./cards.js');
+const { sendUsers, doesUserExist, sendUser } = require('./users.js');
+
+const noSuchAddress = (req, res) => {
+	res.status(404);
+  res.send({ "message": "Запрашиваемый ресурс не найден" });
+};
 
 router.get('/users', sendUsers);
 router.get('/cards', sendCards);
